+++ ZUtil/src/arithmetic.cpp
@@ -8,8 +8,8 @@ double VectorDotProduct( const std::vector<double>& v1, const std::vector<double
 		throw std::logic_error("incompatible eigenfunctions!");
 
 	double sum(0);
-	int size = v1.size();
-	for (int i = 0; i < size; ++i)
+	size_t size = v1.size();
+	for (size_t i = 0; i < size; ++i)
 		sum += v1[i] * v2[i];
 
 	return sum;
@@ -20,8 +20,8 @@ double VectorScalarProduct( const std::vector<double>& v1, const std::vector<dou
 	assert(v1.size() == v2.size() && v1.size() == s.size());
 
 	double sum(0);
-	int size = v1.size();
-	for (int i = 0; i < size; ++i)
+	size_t size = v1.size();
+	for (size_t i = 0; i < size; ++i)
 		sum += v1[i] * s[i] * v2[i];
 
 	return sum;
